% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borg_classification_metrics.R
\name{borg_metrics}
\alias{borg_metrics}
\title{Available Classification Metrics}
\usage{
borg_metrics()
}
\value{
Character vector of available classification metrics.
}
\description{
Returns the valid classification metric names.
}
\examples{
borg_metrics()

}
