% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BORG-package.R
\name{borg_options}
\alias{borg_options}
\title{Get Current BORG Options}
\usage{
borg_options()
}
\value{
A named list of current BORG options.
}
\description{
Returns the current state of BORG configuration options.
}
\examples{
borg_options()

}
