% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borg_audit.R
\name{print.borg_cv_report}
\alias{print.borg_cv_report}
\title{Print CV Leakage Report}
\usage{
\method{print}{borg_cv_report}(x, ...)
}
\arguments{
\item{x}{A borg_cv_report object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print CV Leakage Report
}
